slcand: add example scripts for udev/systemd
The slcand daemon must be started after a serial CAN device is plugged
in to create the network interface (e.g. slcan0).  This can be done
manually but it is inconvenient.

Add example scripts for udev and systemd that can be used to start
slcand automatically when plugged in and to stop it when unplugged.

Signed-off-by: Jeremiah Mahler <jmmahler@gmail.com>